+++ src/plugin.rs
@@ -147,16 +147,26 @@ pub struct ToggleActions<A: Actionlike> {
     ///
     /// When this is set to false, all corresponding [`ActionState`]s are released
     pub enabled: bool,
-    _phantom: PhantomData<A>,
+    /// Marker that stores the type of action to toggle
+    pub phantom: PhantomData<A>,
+}
+
+impl<A: Actionlike> ToggleActions<A> {
+    /// A [`ToggleActions`] in enabled state.
+    pub const ENABLED: ToggleActions::<A> = ToggleActions::<A>::new(true);
+    /// A [`ToggleActions`] in disabled state.
+    pub const DISABLED: ToggleActions::<A> = ToggleActions::<A>::new(false);
+
+    /// Constructs a new [`ToggleActions`] with a specific state.
+    pub const fn new(enabled: bool) -> Self {
+        Self { enabled, phantom: PhantomData::<A>, }
+    }
 }
 
 // Implement manually to not require [`Default`] for `A`
 impl<A: Actionlike> Default for ToggleActions<A> {
     fn default() -> Self {
-        Self {
-            enabled: true,
-            _phantom: PhantomData::<A>,
-        }
+        Self::new(true)
     }
 }